inference/recording-replay-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
kafka-python==2.0.2
minio==7.2.5
opencv-python-headless==4.9.0.80
structlog==24.1.0
python-multipart==0.0.9
